Refactor, add support for static functions
