+++ Sources/Macros/Internal/Extensions/AwaitExprSyntax+Extension.swift
@@ -1,0 +1,7 @@
+import SwiftSyntax
+
+extension AwaitExprSyntax {
+  init(_ expression: FunctionCallExprSyntax) {
+    self.init(expression: expression)
+  }
+}
