+++ Sources/Macros/Internal/Extensions/FunctionCallExprSyntax+Extension.swift
@@ -1,0 +1,8 @@
+import SwiftSyntax
+
+extension FunctionCallExprSyntax {
+  var readable: String {
+    self.trimmedDescription
+  }
+}
+
