+++ Sources/Macros/Internal/Extensions/FunctionSignatureSyntax+Extension.swift
@@ -1,0 +1,10 @@
+import SwiftSyntax
+
+extension FunctionSignatureSyntax {
+  var simplifySignature: FunctionSignatureSyntax {
+    return self.modify { signature in
+      signature.parameterClause = signature.parameterClause.simplifyParameters
+      return signature
+    }
+  }
+}
