+++ Sources/Macros/Internal/Extensions/LabeledExprListSyntaxElement+Extension.swift
@@ -1,0 +1,7 @@
+import SwiftSyntax
+
+extension LabeledExprListSyntax.Element {
+  func parameter(is tokenKind: TokenKind) -> Bool {
+    self.label?.tokenKind == tokenKind
+  }
+}
