+++ Sources/Macros/Internal/Extensions/SyntaxProtocol+Extension.swift
@@ -1,10 +1,6 @@
 import SwiftSyntax
 
 extension SyntaxProtocol {
-  mutating func mutate(_ build: @Sendable @escaping (inout Self) -> Self) -> Self {
-    build(&self)
-  }
-  
   func modify(_ build: @Sendable @escaping (inout Self) -> Self) -> Self {
     var copy = self
     return build(&copy)
