+++ Sources/Macros/Internal/Extensions/TokenKind+Extension.swift
@@ -1,0 +1,6 @@
+import SwiftSyntax
+
+extension TokenKind {
+  static let using = TokenKind.identifier("using")
+  static let omit = TokenKind.identifier("Omit")
+}
