+++ Sources/Macros/Internal/Extensions/TokenSyntax+Extension.swift
@@ -1,0 +1,10 @@
+import SwiftSyntax
+
+extension TokenSyntax {
+  static let log = TokenSyntax.identifier("Log")
+  static let loggable = TokenSyntax.identifier("Loggable")
+  static let result = TokenSyntax.identifier("result")
+  static let error = TokenSyntax.identifier("error")
+  static let location = TokenSyntax.identifier("location")
+  static let `default` = TokenSyntax.identifier("default")
+}
