+++ Sources/Macros/Internal/Extensions/TryExprSyntax+Extension.swift
@@ -1,0 +1,12 @@
+import SwiftSyntax
+
+extension TryExprSyntax {
+  init(_ expression: FunctionCallExprSyntax) {
+    self.init(expression: expression)
+  }
+  
+  init(_ expression: AwaitExprSyntax) {
+    self.init(expression: expression)
+  }
+}
+
