+++ Sources/Macros/Logged+MemberAttributeMacro.swift
@@ -1,0 +1,17 @@
+import SwiftSyntax
+import SwiftSyntaxBuilder
+import SwiftSyntaxMacros
+
+public struct LoggedMacro: MemberAttributeMacro {
+  public static func expansion(
+    of node: AttributeSyntax,
+    attachedTo declaration: some DeclGroupSyntax,
+    providingAttributesFor member: some DeclSyntaxProtocol,
+    in context: some MacroExpansionContext
+  ) throws -> [AttributeSyntax] {
+    guard let member = FunctionDeclSyntax(member),
+          !member.attributes.contains(where: \.isOmit)
+    else { return [] }
+    return [AttributeSyntax.copy(node)]
+  }
+}
