+++ Sources/Macros/Omit+BodyMacro.swift
@@ -1,11 +1,12 @@
 import SwiftSyntax
-import SwiftSyntaxBuilder
 import SwiftSyntaxMacros
 
-public struct OmitMacro: BodyMacro {
+public struct OmitMacro: BodyMacro, BodyMacroBuilder {
+  typealias Body = [CodeBlockItemSyntax]
+  
   public static func expansion(
     of node: AttributeSyntax,
     providingBodyFor declaration: some DeclSyntaxProtocol & WithOptionalCodeBlockSyntax,
     in context: some MacroExpansionContext
-  ) throws -> [CodeBlockItemSyntax] { return [] }
+  ) throws -> [CodeBlockItemSyntax] { body() }
 }
